MediPriceInsight Project/Price Transparency Report/requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0 